fix: decouple playground with risedev (#7249) (#7309)
Previously, the `risingwave playground` command trys to load `risedev.yml` to get the profile for playground. It fallbacks to default configurations if `risedev.yml` is not found (Note: users using pre-built binary always go in this path). This can cause confusion because the configuration will be different in different working directory and even bugs like #7238 (Fix #7238)

This PR makes `playground` self-contained, so that users can run the `risingwave` binary with the same configuration in any directory. The code for playground is also more straitforward and easier to change for developers.

Other minor changes about docker:
- Add a check to ensure the image starts correctly
- Remove `docker/build.sh`, which was used to build individual binaries. They are deprecated after #6653
- Fix `docker-compose.yml`: use latest image
- Add some docs about docker

Approved-By: BugenZhao
Approved-By: huangjw806
Approved-By: fuyufjh

Co-authored-by: xxchan <xxchan22f@gmail.com>